engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
playwright-stealth==2.0.0
pydantic==2.5.0
requests==2.31.0
httpx==0.27.0
beautifulsoup4==4.12.3
google-generativeai==0.8.0
openai==1.55.0
firebase-admin==6.6.0
google-cloud-secret-manager==2.20.0
google-cloud-firestore==2.19.0
python-dotenv==1.0.1